BLE: create ble observer service

+++ ble/src/main/kotlin/br/ufc/predetect/ble/filters/RSSIFilter.kt
@@ -2,6 +2,6 @@ package br.ufc.predetect.ble.filters
 
 interface RSSIFilter {
 
-    fun filter(advertisingPackets: List<Int>) : Int
+    fun filter(advertisingPackets: List<Int>) : Double
 
 }
